Addressed Ed's feedback
